access check and time zone (#4)
* access check and time zone

*  Setting timezone in a function.
* Changed start date and end date types from String to Number
* Checking access to the database on all requests.
* Showing a solution to a user, in case he doesn't have access to a database.

* Made the suggested changes

*  Created the functions to avoid code repetition.
*  Application of them in their proper places.

+++ src/components/BodyTableRemoveModel/index.tsx
@@ -5,6 +5,7 @@ import { IoMdClose } from "react-icons/io";
 
 import { useInfoContext } from "../../contexts/userInfoContext";
 import { deleteModelDatabase } from "../../database/firebase";
+import { checkAccess } from "../../helpers/authorizations";
 
 type Props = {
   item: ProductDatabase;
@@ -22,6 +23,11 @@ function BodyTableRemoveModel(props: Props) {
     const user = state.infoUser?.email;
     const token = await state.infoUser?.getIdToken();
     const authorizedDatabase = state.databaseAuth;
+
+    const accessAuthorized = await checkAccess(state);
+    if (!accessAuthorized) {
+      return;
+    }
 
     await deleteModelDatabase(item.id, user, token, authorizedDatabase);
     getProducts();
